Move gen.js to top level
There is no reason to have index.js just import then export gen.js.
It only made sense to do that before when I was combining several
submodules into one export. But that level of indirection is unnecessary.
